Added CRUD repository using DDD. Added comments to all methods

+++ src/main/java/com/example/ea_java_2/models/CustomerSpender.java
@@ -1,9 +1,12 @@
 package com.example.ea_java_2.models;
 
+/**
+ * Contains information about the customers spending.
+ */
 public record CustomerSpender(
         int customerId,
         String name,
-        int totalSpent
+        float totalSpent
 ) {
 
 }
